fixed various bugs
