+++ saltie.py
@@ -156,6 +156,7 @@ class Agent:
             action = self.actions_handler.get_random_option()
         self.previous_action = action
         controller_selection = self.actions_handler.create_controller_from_selection(action)
+        controller_selection = [max(-1, min(1, control)) for control in controller_selection]
         return controller_selection
 
     def create_model_hash(self):
